Updated proto package to be com.github.dozermapper.protobuf
