--- proto/src/main/java/org/dozer/builder/ProtoBeanBuilder.java
@@ -1,45 +1,0 @@
-/*
- * Copyright 2005-2017 Dozer Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.dozer.builder;
-
-import com.google.protobuf.Message;
-
-import org.dozer.BeanBuilder;
-
-/**
- * @author Dmitry Spikhalskiy
- */
-public class ProtoBeanBuilder implements BeanBuilder {
-  protected Message.Builder internalProtoBuilder;
-  protected Class<? extends Message> beanClass;
-
-  public ProtoBeanBuilder(Message.Builder internalProtoBuilder, Class<? extends Message> beanClass) {
-    this.internalProtoBuilder = internalProtoBuilder;
-    this.beanClass = beanClass;
-  }
-
-  public Class<? extends Message> beanClass() {
-    return beanClass;
-  }
-
-  public Message.Builder internalProtoBuilder() {
-    return internalProtoBuilder;
-  }
-
-  public Object build() {
-    return internalProtoBuilder.build();
-  }
-}
