+++ proto/src/main/resources/META-INF/services/org.dozer.DozerModule
@@ -1,1 +1,1 @@
-org.dozer.ProtobufSupportModule
+com.github.dozermapper.protobuf.ProtobufSupportModule
