+++ proto/src/test/java/com/github/dozermapper/protobuf/vo/proto/MapExample.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2005-2017 Dozer Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.dozermapper.protobuf.vo.proto;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class MapExample {
+    public Map<String, String> getValue() {
+        return value;
+    }
+
+    public void setValue(Map<String, String> value) {
+        this.value = value;
+    }
+
+    private Map<String,String> value = new HashMap<>();
+
+    public void put(String key, String value) {
+        this.value.put(key, value);
+    }
+}
