+++ proto/src/test/java/com/github/dozermapper/protobuf/vo/proto/TestObject.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2005-2017 Dozer Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.dozermapper.protobuf.vo.proto;
+
+/**
+ * @author garsombke.franz
+ * @author sullins.ben
+ * @author tierney.matt
+ *
+ */
+public class TestObject {
+
+    private String one;
+    private Integer two;
+
+    public String getOne() {
+        return one;
+    }
+
+    public void setOne(String one) {
+        this.one = one;
+    }
+
+    public Integer getTwo() {
+        return two;
+    }
+
+    public void setTwo(Integer two) {
+        this.two = two;
+    }
+}
