--- proto/src/test/java/org/dozer/vo/proto/LiteTestObjectContainer.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2005-2017 Dozer Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.dozer.vo.proto;
-
-/**
- * @author Dmitry Spikhalskiy
- * @since 05.01.13
- */
-public class LiteTestObjectContainer {
-
-    private LiteTestObject object;
-
-    public LiteTestObject getObject() {
-        return object;
-    }
-
-    public void setObject(LiteTestObject object) {
-        this.object = object;
-    }
-}
